Refactor chacha cuda to be able to test cuda crate but not in OpenCL
chacha not implemeted in OpenCL
