+++ Cargo.toml
@@ -4,6 +4,8 @@ members = [
     "bench-streamer",
     "bench-tps",
     "banking-bench",
+    "chacha",
+    "chacha-cuda",
     "chacha-sys",
     "client",
     "core",
@@ -38,6 +40,8 @@ members = [
     "programs/vest",
     "programs/vote",
     "archiver",
+    "archiver-lib",
+    "archiver-utils",
     "runtime",
     "sdk",
     "sdk-c",
