+++ archiver-lib/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "solana-archiver-lib"
+version = "0.23.0"
+description = "Solana Archiver Library"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+edition = "2018"
+
+[dependencies]
+bincode = "1.2.1"
+crossbeam-channel = "0.3"
+ed25519-dalek = "=1.0.0-pre.1"
+log = "0.4.8"
+rand = "0.6.5"
+rand_chacha = "0.1.1"
+solana-client = { path = "../client", version = "0.23.0" }
+solana-storage-program = { path = "../programs/storage", version = "0.23.0" }
+serde = "1.0.104"
+serde_json = "1.0.44"
+serde_derive = "1.0.103"
+solana-net-utils = { path = "../net-utils", version = "0.23.0" }
+solana-chacha = { path = "../chacha", version = "0.23.0" }
+solana-chacha-sys = { path = "../chacha-sys", version = "0.23.0" }
+solana-ledger = { path = "../ledger", version = "0.23.0" }
+solana-logger = { path = "../logger", version = "0.23.0" }
+solana-perf = { path = "../perf", version = "0.23.0" }
+solana-sdk = { path = "../sdk", version = "0.23.0" }
+solana-core = { path = "../core", version = "0.23.0" }
+solana-archiver-utils = { path = "../archiver-utils", version = "0.23.0" }
+solana-metrics = { path = "../metrics", version = "0.23.0" }
+
+[dev-dependencies]
+hex = "0.4.0"
+
+[lib]
+name = "solana_archiver_lib"
