+++ archiver/Cargo.toml
@@ -14,6 +14,7 @@ solana-clap-utils = { path = "../clap-utils", version = "0.23.0" }
 solana-core = { path = "../core", version = "0.23.0" }
 solana-logger = { path = "../logger", version = "0.23.0" }
 solana-metrics = { path = "../metrics", version = "0.23.0" }
+solana-archiver-lib = { path = "../archiver-lib", version = "0.23.0" }
 solana-net-utils = { path = "../net-utils", version = "0.23.0" }
 solana-sdk = { path = "../sdk", version = "0.23.0" }
