+++ chacha/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "solana-chacha"
+version = "0.23.0"
+description = "Solana Chacha APIs"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+edition = "2018"
+
+[dependencies]
+log = "0.4.8"
+rand = "0.6.5"
+rand_chacha = "0.1.1"
+solana-chacha-sys = { path = "../chacha-sys", version = "0.23.0" }
+solana-ledger = { path = "../ledger", version = "0.23.0" }
+solana-logger = { path = "../logger", version = "0.23.0" }
+solana-perf = { path = "../perf", version = "0.23.0" }
+solana-sdk = { path = "../sdk", version = "0.23.0" }
+
+[dev-dependencies]
+hex-literal = "0.2.1"
+
+[lib]
+name = "solana_chacha"
