+++ core/Cargo.toml
@@ -43,7 +43,6 @@ serde_derive = "1.0.103"
 serde_json = "1.0.44"
 solana-budget-program = { path = "../programs/budget", version = "0.23.0" }
 solana-clap-utils = { path = "../clap-utils", version = "0.23.0" }
-solana-chacha-sys = { path = "../chacha-sys", version = "0.23.0" }
 solana-client = { path = "../client", version = "0.23.0" }
 solana-faucet = { path = "../faucet", version = "0.23.0" }
 ed25519-dalek = "=1.0.0-pre.1"
@@ -53,6 +52,7 @@ solana-merkle-tree = { path = "../merkle-tree", version = "0.23.0" }
 solana-metrics = { path = "../metrics", version = "0.23.0" }
 solana-measure = { path = "../measure", version = "0.23.0" }
 solana-net-utils = { path = "../net-utils", version = "0.23.0" }
+solana-chacha-cuda = { path = "../chacha-cuda", version = "0.23.0" }
 solana-perf = { path = "../perf", version = "0.23.0" }
 solana-runtime = { path = "../runtime", version = "0.23.0" }
 solana-sdk = { path = "../sdk", version = "0.23.0" }
@@ -74,7 +74,6 @@ solana-rayon-threadlimit = { path = "../rayon-threadlimit", version = "0.23.0" }
 reed-solomon-erasure = { package = "solana-reed-solomon-erasure", version = "4.0.1-3", features = ["simd-accel"] }
 
 [dev-dependencies]
-hex-literal = "0.2.1"
 matches = "0.1.6"
 reqwest = { version = "0.10.1", default-features = false, features = ["blocking", "rustls-tls"] }
 serial_test = "0.3.2"
