+++ local-cluster/src/local_cluster.rs
@@ -1,9 +1,9 @@
 use crate::cluster::{Cluster, ClusterValidatorInfo, ValidatorInfo};
 use itertools::izip;
 use log::*;
+use solana_archiver_lib::archiver::Archiver;
 use solana_client::thin_client::{create_client, ThinClient};
 use solana_core::{
-    archiver::Archiver,
     cluster_info::{Node, VALIDATOR_PORT_RANGE},
     contact_info::ContactInfo,
     genesis_utils::{create_genesis_config_with_leader, GenesisConfigInfo},
